refactor(pipeline): extract guardrails_test.go from engine_test.go [#285]
Move 33 guardrail-related test functions from engine_test.go into
guardrails_test.go, co-locating tests with the code they exercise
(checkBudget, checkPhaseBudget, gatePhase, gateRework, gateVerifyFail,
handlePatchExhausted, gatePatchResult, downgradeToFollowUps,
detectRegression, extractFailingCriteria). Integration tests that
exercise the full pipeline loop remain in engine_test.go.
